# api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: FinDoctor Client API
  version: 0.1.0
  summary: Client-facing API for financial accounts, ledger transactions, assets, liabilities, goals, tags, and analytics.
  license:
    name: Proprietary
    url: https://findoctor.example/legal/api-license
  description: |
    FinDoctor exposes user-owned financial data as explicit CRUD resources and keeps reference/admin data separate.

    Transfers are modeled as business objects connected to two underlying ledger transactions:
    one debit transaction on the source account and one credit transaction on the destination account.
    Creating, updating, or deleting a transfer must happen transactionally with those ledger rows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.findoctor.example
    description: Production placeholder
security:
  - bearerAuth: []
...
paths:
...
  /reference/categories:
    get:
      tags: [Reference Data]
      summary: List transaction categories
      operationId: listCategories
      parameters:
        - $ref: "#/components/parameters/CategoryTypeQuery"
        - $ref: "#/components/parameters/ParentIdQuery"
      responses:
        "200":
          description: Category tree or filtered category list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryList"
    post:
      tags: [Reference Data]
      summary: Create transaction category
      description: Categories are global in the current schema. If per-user categories are needed, add `user_id` before exposing this broadly.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "422":
          $ref: "#/components/responses/ValidationError"
  /reference/categories/{categoryId}:
    get:
      tags: [Reference Data]
      summary: Get transaction category
      operationId: getCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Reference Data]
      summary: Update transaction category
      operationId: updateCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Reference Data]
      summary: Delete transaction category
      operationId: deleteCategory
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted
        "409":
          $ref: "#/components/responses/Conflict"
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/DateTimeFromQuery"
        - $ref: "#/components/parameters/DateTimeToQuery"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/TransactionType"
        - name: merchantId
          in: query
          schema:
            type: string
            format: uuid
        - name: recurringTransactionId
          in: query
          schema:
            type: string
            format: uuid
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
    post:
      tags: [Transactions]
      summary: Create transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreate"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /transactions/{transactionId}:
    get:
      tags: [Transactions]
      summary: Get transaction
      operationId: getTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Transactions]
      summary: Update transaction
      operationId: updateTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Transactions]
      summary: Delete transaction
      operationId: deleteTransaction
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "204":
          description: Transaction deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /transactions/{transactionId}/tags:
    get:
      tags: [Transactions, Tags]
      summary: List transaction tags
      operationId: listTransactionTags
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    put:
      tags: [Transactions, Tags]
      summary: Replace all tags on a transaction
      operationId: replaceTransactionTags
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tagIds]
              properties:
                tagIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  uniqueItems: true
      responses:
        "200":
          description: Updated transaction tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
  /transactions/{transactionId}/tags/{tagId}:
    post:
      tags: [Transactions, Tags]
      summary: Attach tag to transaction
      operationId: addTransactionTag
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag attached
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Transactions, Tags]
      summary: Detach tag from transaction
      operationId: removeTransactionTag
      parameters:
        - $ref: "#/components/parameters/TransactionId"
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag detached
        "404":
          $ref: "#/components/responses/NotFound"
...
  /recurring-transactions:
    get:
      tags: [Recurring Transactions]
      summary: List recurring transaction patterns
      operationId: listRecurringTransactions
      parameters:
        - $ref: "#/components/parameters/ActiveQuery"
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: liabilityId
          in: query
          schema:
            type: string
            format: uuid
        - name: nextPaymentBefore
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Recurring transaction patterns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransactionList"
    post:
      tags: [Recurring Transactions]
      summary: Create recurring transaction pattern
      operationId: createRecurringTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringTransactionCreate"
      responses:
        "201":
          description: Created recurring pattern
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
  /recurring-transactions/{recurringTransactionId}:
    get:
      tags: [Recurring Transactions]
      summary: Get recurring transaction pattern
      operationId: getRecurringTransaction
      parameters:
        - $ref: "#/components/parameters/RecurringTransactionId"
      responses:
        "200":
          description: Recurring pattern
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Recurring Transactions]
      summary: Update recurring transaction pattern
      operationId: updateRecurringTransaction
      parameters:
        - $ref: "#/components/parameters/RecurringTransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringTransactionUpdate"
      responses:
        "200":
          description: Updated recurring pattern
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringTransaction"
    delete:
      tags: [Recurring Transactions]
      summary: Deactivate recurring transaction pattern
      operationId: deleteRecurringTransaction
      parameters:
        - $ref: "#/components/parameters/RecurringTransactionId"
      responses:
        "204":
          description: Recurring pattern deactivated
...
  /tags:
    get:
      tags: [Tags]
      summary: List user tags
      operationId: listTags
      responses:
        "200":
          description: User tags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      tags: [Tags]
      summary: Create tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagCreate"
      responses:
        "201":
          description: Created tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "409":
          $ref: "#/components/responses/Conflict"
  /tags/{tagId}:
    get:
      tags: [Tags]
      summary: Get tag
      operationId: getTag
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "200":
          description: Tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    patch:
      tags: [Tags]
      summary: Update tag
      operationId: updateTag
      parameters:
        - $ref: "#/components/parameters/TagId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagUpdate"
      responses:
        "200":
          description: Updated tag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
    delete:
      tags: [Tags]
      summary: Delete tag
      operationId: deleteTag
      parameters:
        - $ref: "#/components/parameters/TagId"
      responses:
        "204":
          description: Tag deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CategoryTypeQuery:
      name: type
      in: query
      schema:
        $ref: "#/components/schemas/CategoryType"
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DateTimeFromQuery:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    DateTimeToQuery:
      name: to
      in: query
      schema:
        type: string
        format: date-time
...
    ParentIdQuery:
      name: parentId
      in: query
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 250
        default: 50
    RecurringTransactionId:
      name: recurringTransactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ActiveQuery:
      name: isActive
      in: query
      schema:
        type: boolean
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or uniqueness constraints
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
  schemas:
...
    Category:
      type: object
      required: [id, type, name]
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/CategoryType"
        name:
          type: string
        description:
          type: [string, "null"]
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    CategoryCreate:
      type: object
      required: [type, name]
      properties:
        parentId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/CategoryType"
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, "null"]
    CategoryUpdate:
      type: object
      properties:
        parentId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/CategoryType"
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, "null"]
    CategoryList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
    CategoryType:
      type: string
      enum: [income, expense]
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
      pattern: "^[A-Z]{3}$"
      examples: [USD, EUR, RUB]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
...
    MoneyAmount:
      type: string
      pattern: "^-?\\d+(\\.\\d{1,8})?$"
      description: Decimal amount encoded as a string to avoid floating point loss.
      examples: ["1250.00"]
    MoneyAmountNullable:
      type: [string, "null"]
      pattern: "^-?\\d+(\\.\\d{1,8})?$"
...
    PageMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
...
    RecurringFrequency:
      type: string
      enum: [daily, weekly, monthly, yearly]
    RecurringOperationType:
      type: string
      enum: [income, expense, transfer, liability_payment]
    RecurringTransaction:
      allOf:
        - $ref: "#/components/schemas/RecurringTransactionCreate"
        - type: object
          required: [id, userId, autoGenerated, isActive]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            autoGenerated:
              type: boolean
            confidenceScore:
              type: [number, "null"]
            isActive:
              type: boolean
    RecurringTransactionCreate:
      type: object
      required: [accountId, operationType, name, frequency]
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        liabilityId:
          type: [string, "null"]
          format: uuid
        operationType:
          $ref: "#/components/schemas/RecurringOperationType"
        name:
          type: string
          minLength: 1
          maxLength: 200
        expectedAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currency:
          oneOf:
            - $ref: "#/components/schemas/CurrencyCode"
            - type: "null"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
        intervalCount:
          type: integer
          minimum: 1
          default: 1
        dayOfMonth:
          type: [integer, "null"]
          minimum: 1
          maximum: 31
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
        nextPaymentDate:
          type: [string, "null"]
          format: date
    RecurringTransactionUpdate:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        liabilityId:
          type: [string, "null"]
          format: uuid
        operationType:
          $ref: "#/components/schemas/RecurringOperationType"
        name:
          type: string
          minLength: 1
          maxLength: 200
        expectedAmount:
          $ref: "#/components/schemas/MoneyAmountNullable"
        currency:
          oneOf:
            - $ref: "#/components/schemas/CurrencyCode"
            - type: "null"
        frequency:
          $ref: "#/components/schemas/RecurringFrequency"
        intervalCount:
          type: integer
          minimum: 1
        dayOfMonth:
          type: [integer, "null"]
          minimum: 1
          maximum: 31
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
        nextPaymentDate:
          type: [string, "null"]
          format: date
        isActive:
          type: boolean
    RecurringTransactionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecurringTransaction"
...
    Tag:
      type: object
      required: [id, userId, name]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
    TagCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    TagUpdate:
      allOf:
        - $ref: "#/components/schemas/TagCreate"
    TagList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
    Transaction:
      allOf:
        - $ref: "#/components/schemas/TransactionCreate"
        - type: object
          required: [id, userId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            transferId:
              type: [string, "null"]
              format: uuid
              description: Present when this ledger row belongs to a transfer.
            transferLeg:
              type: [string, "null"]
              enum: [debit, credit, null]
            createdAt:
              type: string
              format: date-time
            tags:
              type: array
              items:
                $ref: "#/components/schemas/Tag"
    TransactionCreate:
      type: object
      required: [accountId, type, amount, currency, transactionDatetime]
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        merchantId:
          type: [string, "null"]
          format: uuid
        merchantName:
          type: [string, "null"]
        geoLocation:
          type: [string, "null"]
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        externalId:
          type: [string, "null"]
        tagIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
...
    TransactionPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        meta:
          $ref: "#/components/schemas/PageMeta"
    TransactionType:
      type: string
      enum: [income, expense, transfer]
    TransactionUpdate:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        transactionDatetime:
          type: string
          format: date-time
        description:
          type: [string, "null"]
        merchantId:
          type: [string, "null"]
          format: uuid
        merchantName:
          type: [string, "null"]
        geoLocation:
          type: [string, "null"]
        recurringTransactionId:
          type: [string, "null"]
          format: uuid
        externalId:
          type: [string, "null"]
...